+++ runtime/binding-http/src/main/java/io/aklivity/zilla/runtime/binding/http/config/HttpOptionsConfig.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package io.aklivity.zilla.runtime.binding.http.internal.config;
+package io.aklivity.zilla.runtime.binding.http.config;
 
 import java.util.Map;
 import java.util.SortedSet;
