+++ runtime/binding-http/src/main/java/io/aklivity/zilla/runtime/binding/http/internal/config/HttpConditionMatcher.java
@@ -21,6 +21,8 @@ import java.util.function.Function;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import io.aklivity.zilla.runtime.binding.http.config.HttpConditionConfig;
+
 public final class HttpConditionMatcher
 {
     private final Map<String, Matcher> headersMatch;
