+++ runtime/binding-http/src/test/java/io/aklivity/zilla/runtime/binding/http/internal/config/HttpConditionConfigAdapterTest.java
@@ -28,6 +28,8 @@ import jakarta.json.bind.JsonbConfig;
 import org.junit.Before;
 import org.junit.Test;
 
+import io.aklivity.zilla.runtime.binding.http.config.HttpConditionConfig;
+
 public class HttpConditionConfigAdapterTest
 {
     private Jsonb jsonb;
