+++ runtime/binding-tcp/src/main/moditect/module-info.java
@@ -17,6 +17,8 @@ module io.aklivity.zilla.runtime.binding.tcp
 {
     requires io.aklivity.zilla.runtime.engine;
 
+    exports io.aklivity.zilla.runtime.binding.tcp.config;
+
     provides io.aklivity.zilla.runtime.engine.binding.BindingFactorySpi
         with io.aklivity.zilla.runtime.binding.tcp.internal.TcpBindingFactorySpi;
