+++ runtime/binding-tls/src/main/moditect/module-info.java
@@ -19,6 +19,8 @@ module io.aklivity.zilla.runtime.binding.tls
 
     requires io.aklivity.zilla.runtime.engine;
 
+    exports io.aklivity.zilla.runtime.binding.tls.config;
+
     provides io.aklivity.zilla.runtime.engine.binding.BindingFactorySpi
         with io.aklivity.zilla.runtime.binding.tls.internal.TlsBindingFactorySpi;
